pyproject.toml
[tool.poetry]
name = "product-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
sqlmodel = "^0.0.21"
uvicorn = "^0.30.5"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
